move user context to login component
