+++ src/components/app/App.tsx
@@ -7,15 +7,8 @@ import * as rest from "../../rest/interface";
 import '../../styles.css';
 import { Admin } from "../admin/Admin";
 import ConfigDrop from '../configdrop/ConfigDrop';
-import { Login } from "../login/Login";
+import { Login, UserContextConsumer } from "../login/Login";
 import { PatchContainer } from '../patch/PatchContainer';
-
-export const UserContext = React.createContext({
-  username: "",
-});
-console.log("setting user context")
-export const UserContextProvider = UserContext.Provider;
-export const UserContextConsumer = UserContext.Consumer;
 
 interface State {
   APIClient: rest.Evergreen;
@@ -39,45 +32,51 @@ export class Evergreen extends React.Component<Props, State> {
 
   public render() {
     const admin = () => <Admin APIClient={this.state.APIClient} />
-    const patches = () => <PatchContainer client={this.state.APIClient} username={this.state.username} onFinishStateUpdate={null}/>
-    const config = () => <ConfigDrop updateClientConfig={this.updateConfig} onLoadFinished={null}/>
+    const patches = () => <PatchContainer client={this.state.APIClient} username={this.state.username} onFinishStateUpdate={null} />
+    const config = () => <ConfigDrop updateClientConfig={this.updateConfig} onLoadFinished={null} />
     const menuOpen = Boolean(this.state.MenuAnchor);
 
     return (
       <div className="App">
-        <HashRouter>
-          <AppBar position="fixed" className="app-bar">
-            <Toolbar>
-              <Typography variant="h5" color="inherit" noWrap={true}>
-                Evergreen
-              </Typography>
-              <div className="spacer" /> 
-              <IconButton className="menu" color="inherit" id="mainAppIcon" onClick={this.openMenu}>
-                <MenuIcon.default />
-              </IconButton>
-              <Menu id="mainAppMenu" open={menuOpen} anchorEl={this.state.MenuAnchor} 
-                anchorOrigin={{vertical: 'top', horizontal: 'right',}}
-                transformOrigin={{vertical: 'top', horizontal: 'right',}}
-                onClose={this.closeMenu}>
-                <MenuItem onClick={this.closeMenu}>
-                  <NavLink to="/admin"> Admin Page</NavLink>
-                </MenuItem>
-                <MenuItem onClick={this.closeMenu}>
-                  <NavLink to="/patches">My Patches</NavLink> 
-                </MenuItem>           
-                <MenuItem onClick={this.closeMenu}>
-                  <NavLink to="/config">Upload Config File</NavLink> 
-                </MenuItem>
-              </Menu>
-              <Login client={this.state.APIClient} updateUsername={this.updateUsername}/>
-            </Toolbar>
-          </AppBar>
-          <div className="App-intro">
-            <Route path="/admin" render={admin} />
-            <Route path="/config" render={config} />
-            <Route path="/patches" render={patches} />
-          </div>
-        </HashRouter>
+        <UserContextConsumer>
+          {props => {
+            return (
+              <HashRouter>
+                <AppBar position="fixed" className="app-bar">
+                  <Toolbar>
+                    <Typography variant="h5" color="inherit" noWrap={true}>
+                      Evergreen
+                    </Typography>
+                    <div className="spacer" />
+                    <IconButton className="menu" color="inherit" id="mainAppIcon" onClick={this.openMenu}>
+                      <MenuIcon.default />
+                    </IconButton>
+                    <Menu id="mainAppMenu" open={menuOpen} anchorEl={this.state.MenuAnchor}
+                      anchorOrigin={{ vertical: 'top', horizontal: 'right', }}
+                      transformOrigin={{ vertical: 'top', horizontal: 'right', }}
+                      onClose={this.closeMenu}>
+                      <MenuItem onClick={this.closeMenu}>
+                        <NavLink to="/admin"> Admin Page</NavLink>
+                      </MenuItem>
+                      <MenuItem onClick={this.closeMenu}>
+                        <NavLink to="/patches">My Patches</NavLink>
+                      </MenuItem>
+                      <MenuItem onClick={this.closeMenu}>
+                        <NavLink to="/config">Upload Config File</NavLink>
+                      </MenuItem>
+                    </Menu>
+                    <Login client={this.state.APIClient} updateUsername={this.updateUsername} />
+                  </Toolbar>
+                </AppBar>
+                <div className="App-intro">
+                  <Route path="/admin" render={admin} />
+                  <Route path="/config" render={config} />
+                  <Route path="/patches" render={patches} />
+                </div>
+              </HashRouter>
+            )
+          }}
+        </UserContextConsumer>
       </div>
     );
   }
@@ -104,15 +103,15 @@ export class Evergreen extends React.Component<Props, State> {
 
   private tryLoadConfig = () => {
     fetch(configPath).then((resp: Response) => {
-        resp.json().then( (config: object) => {
-          if (IsValidConfig(config)) {
-            this.updateConfig(config as ClientConfig);
-          } else {
-            console.log("Config is missing required fields");
-          }
-        }, (reason: any) => {
-          console.log("Error parsing config. You may need to manually drop a config file. Error: " + reason);
-        });
+      resp.json().then((config: object) => {
+        if (IsValidConfig(config)) {
+          this.updateConfig(config as ClientConfig);
+        } else {
+          console.log("Config is missing required fields");
+        }
+      }, (reason: any) => {
+        console.log("Error parsing config. You may need to manually drop a config file. Error: " + reason);
+      });
     });
   }
 }
